Move source and config pkgs into a common internal dir.
