+++ internal/source/source_test.go
@@ -21,7 +21,7 @@ import (
 
 	"gopkg.in/yaml.v3"
 
-	"github.com/ossf/malicious-packages/cmd/ingest/source"
+	"github.com/ossf/malicious-packages/internal/source"
 )
 
 func TestParseSources_InvalidIDs(t *testing.T) {
